+++ src/main/java/com/nhnacademy/order_server/adapter/CartClient.java
@@ -1,7 +1,8 @@
 package com.nhnacademy.order_server.adapter;
 
 import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.RequestHeader;
 
 @FeignClient(name = "TEAM5-MEMBER-SERVER", contextId = "cartClient")
 public interface CartClient {
