+++ src/main/java/com/nhnacademy/order_server/dto/response/CouponCalculationResponse.java
@@ -1,8 +1,8 @@
 package com.nhnacademy.order_server.dto.response;
 
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
 
 @Getter
 @NoArgsConstructor
