+++ src/main/java/com/nhnacademy/order_server/dto/response/MemberCouponResponse.java
@@ -1,11 +1,10 @@
 package com.nhnacademy.order_server.dto.response;
 
+import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
 
 @Getter
 @Builder
