+++ src/main/java/com/nhnacademy/order_server/dto/response/OrderValidationInfoResponse.java
@@ -1,8 +1,6 @@
 package com.nhnacademy.order_server.dto.response;
 
 import com.nhnacademy.order_server.entity.Order;
-import com.nhnacademy.order_server.exception.OrderErrorCode;
-import com.nhnacademy.order_server.exception.OrderException;
 import lombok.Builder;
 import lombok.Getter;
