+++ src/main/java/com/nhnacademy/order_server/exception/ErrorResponse.java
@@ -1,8 +1,8 @@
 package com.nhnacademy.order_server.exception;
 
+import java.time.LocalDateTime;
 import lombok.Builder;
 import lombok.Getter;
-import java.time.LocalDateTime;
 
 @Getter
 @Builder
