+++ src/main/java/com/nhnacademy/order_server/repository/DeliveryPolicyRepository.java
@@ -1,11 +1,8 @@
 package com.nhnacademy.order_server.repository;
 
 import com.nhnacademy.order_server.entity.DeliveryPolicy;
-import jakarta.persistence.LockModeType;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Lock;
-
 import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DeliveryPolicyRepository extends JpaRepository<DeliveryPolicy, Long> {
