+++ src/main/java/com/nhnacademy/order_server/repository/OrderRepository.java
@@ -2,15 +2,14 @@ package com.nhnacademy.order_server.repository;
 
 import com.nhnacademy.order_server.entity.Order;
 import com.nhnacademy.order_server.entity.enums.DeliveryStatus;
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Optional;
 
 public interface OrderRepository extends JpaRepository<Order,Long> {
