+++ src/main/java/com/nhnacademy/order_server/repository/WrapperRepository.java
@@ -1,9 +1,8 @@
 package com.nhnacademy.order_server.repository;
 
 import com.nhnacademy.order_server.entity.Wrapper;
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WrapperRepository extends JpaRepository<Wrapper,Long> {
