+++ src/main/java/com/nhnacademy/order_server/service/DeliveryPolicyService.java
@@ -3,7 +3,6 @@ package com.nhnacademy.order_server.service;
 import com.nhnacademy.order_server.dto.request.DeliveryPolicyRequest;
 import com.nhnacademy.order_server.dto.response.DeliveryPolicyResponse;
 import com.nhnacademy.order_server.entity.DeliveryPolicy;
-
 import java.util.List;
 
 public interface DeliveryPolicyService {
